state library (#246)
* state library

* tests wip

* extract api module

* whitespace

* more whitespace

* avro, core & graphql tests

* integration test

* cleanup deps

* fix typo

* change EventCorrelator contract

* kafka sender tests

* more tests, rework correlator, support deletes

* small puml update

* tidy up

* javadoc

* fix changelog and new module versions

* ApolloExecutor public

* not the inner class

* executor support returning response

* move correlator to an internal package

* remove close from EntityView

* import order

* example

* example copyright headers

* bump apollo & plugin version

* Optional tag

* throw on unexpected type

+++ notification-support/src/test/java/com/expediagroup/streamplatform/streamregistry/core/events/NotificationEventUtilsTest.java
@@ -169,7 +169,7 @@ public class NotificationEventUtilsTest {
     Assert.assertEquals(type, avroEvent.getStreamEntity().getType());
     Assert.assertEquals(configJson, avroEvent.getStreamEntity().getConfigurationString());
     Assert.assertEquals(statusJson, avroEvent.getStreamEntity().getStatusString());
-    Assert.assertEquals(version, avroEvent.getStreamEntity().getVersion().intValue());
+    Assert.assertEquals(version, avroEvent.getStreamEntity().getVersion());
   }
 
   // Don't remove, is loaded by reflection for this test...
